cleanup of main script flow
I had done a lot of ad-hoc training and testing with lyrics by genre- so I cleaned up areas like analysis of Phrases, LDA Topic Modeling, and even the core instantiation of lyrics objects.  

I added functions for common tasks with LDA and Phrases that can now be called in most cases simply by passing an instance of a lyrics object (objects of class MusicalMeg).

I'm refactoring in the direction of being able to do all the various plotting and modeling (doc2vec, word2vec, tfidf and vocab histograms, LDA topic modeling,  and phrases analysis) simply by passing an instance of Musical Meg class.  My intent is for each MusicalMeg class to be a virtual container of lyrics for ALL artists, tracks and lyrics in a particular genre.